+++ app/src/main/AndroidManifest.xml
@@ -21,7 +21,7 @@
         <category android:name="android.intent.category.LAUNCHER" />
       </intent-filter>
     </activity>
-    <activity android:name=".activities.HomeView" />
+    <activity android:name=".views.home.HomeView" />
     <activity android:name=".views.hillfortlist.HillfortListFragment" />
     <activity
       android:name=".views.hillfort.HillfortView"
@@ -29,7 +29,7 @@
     <activity android:name=".views.user.profile.ProfileActivity" />
     <activity android:name=".views.user.settings.SettingsFragment" />
     <activity
-      android:name=".activities.hillfort.HillFortProfileActivity"
+      android:name=".views.hillfortprofile.HillFortProfileActivity"
       android:parentActivityName=".views.hillfortlist.HillfortListFragment" />
 
     <meta-data
