+++ app/src/main/res/layout/activity_home.xml
@@ -5,7 +5,7 @@ xmlns:app="http://schemas.android.com/apk/res-auto"
 xmlns:tools="http://schemas.android.com/tools"
 android:layout_width="match_parent"
 android:layout_height="match_parent"
-tools:context=".activities.HomeView">
+tools:context=".views.home.HomeView">
 
   <android.support.design.widget.AppBarLayout
     android:id="@+id/appBarLayout"
