--- app/src/main/res/layout/activity_maps.xml
@@ -1,9 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<fragment xmlns:android="http://schemas.android.com/apk/res/android"
-  xmlns:map="http://schemas.android.com/apk/res-auto"
-  xmlns:tools="http://schemas.android.com/tools"
-  android:id="@+id/map"
-  android:name="com.google.android.gms.maps.SupportMapFragment"
-  android:layout_width="match_parent"
-  android:layout_height="match_parent"
-  tools:context=".activities.maps.MapsActivity" />
